project is runnning and saving everything

+++ src/main/java/com/devesh/cricket/repository/PlayerMatchStatsRepository.java
@@ -7,6 +7,6 @@ import org.springframework.stereotype.Repository;
 
 
 @Repository
-public interface PlayerMatchStatsRepository extends JpaRepository<PlayerMatchStats, Integer> {
+public interface   PlayerMatchStatsRepository extends JpaRepository<PlayerMatchStats, Long> {
 
 }
